requirements.txt
pandas==1.0.1
requests==2.22.0
torchvision==0.7.0
numpy==1.18.1
opencv_python==4.4.0.42
torch==1.6.0
seaborn==0.10.0
matplotlib==3.1.3
Pillow==7.2.0
scikit_learn==0.23.2
